Started working on the listviews

+++ app/src/main/java/com/example/mobilecoursework/model/CafeItem.kt
@@ -1,4 +1,7 @@
 package com.example.mobilecoursework.model
 
-class CafeItem(var proName: String, var prodPrice: Float,var prodImage: ByteArray ,var prodAvailable: Boolean) {
+class CafeItem(
+    var proName: String, var prodPrice: Double,
+    var prodImage: ByteArray,
+    var prodAvailable: Boolean) {
 }
